chore(build): Configure Gradle to generate JAR in project root
In this commit, I've configured Gradle to generate the JAR file directly in the project's root directory. The JAR is named 'galactic_messenger.jar' and will be placed in the project's root folder.

I've also added a 'clean' task to remove the generated JAR from the root directory when needed.

Signed-off-by: Yassine EL GHERRABI 👮🏽‍♂️ <https://github.com/Yassineelg>