chore(build): Add Main-Class attribute to JAR manifest
In this commit, I've added the 'Main-Class' attribute to the JAR manifest in the build configuration. This is a necessary configuration adjustment to specify the main class that should be executed when launching the JAR with 'java -jar'. The attribute points to 'galactic_messenger.App', which is the main class of our application.

Without this attribute, Java would not know which class to use as the entry point for the JAR, leading to the 'no main manifest attribute' error when trying to run the JAR.

Signed-off-by: Yassine EL GHERRABI 👮🏽‍♂️ <https://github.com/Yassineelg>